Introduce PAUSE/UNPAUSE CLUSTER functionality: A PAUSE CLUSTER command causes that session to take the ClusterLock in exclusive mode. It is obtained after all existing transactions have completed. New transactions will wait till an UNPAUSE CLUSTER command is fired.
PAUSE operation works across transaction boundaries. However if there is an
error, to reduce the risk of deadlocks, the ClusterLock gets released. It is a
superuser only operation.

The intended use case is to carry out NODE metadata updates in isolation. By
pausing new transactions, after the UNPAUSE, they get the latest node pool
information to use for their operations.