Backpatch from PG to provide the post parse analysis hook functionality.
commit a40fa613b516b97c37d87ac1b21fb7aa8a2f2c1b
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Tue Mar 27 15:14:13 2012 -0400

Add a queryId field to Query and PlannedStmt.  This is not used by
the core backend, except for being copied around at appropriate
times.  It's meant to allow plug-ins to track a particular query
forward from parse analysis to execution.

Also, add a post_parse_analyze_hook that gets invoked at the end
of parse analysis (but only for top-level analysis of complete
queries, not cases such as analyzing a domain's default-value
expression).  This is mainly meant to be used to compute and assign
a queryId, but it could have other applications.  core backend,
except for being copied around at appropriate times.  It's meant
to allow plug-ins to track a particular query forward from parse
analysis to execution.