Add options -U (username) and -d (database) in pgxc_monitor
This allows to monitor a node with a simple command using a wanted
user name and/or database for a Coordinator and a Datanode.
In case a password is necessary, it is necessary to provide it
through .pgpass.

Feature design and patch by Nikhil Sontakke.
I just reviewed and corrected a couple of things.