Fix for parameter handling in SPI calls
This commit provides a fix for parameter handling for queries
called through SPI. Most of the issues were related to plpgsql
functions with DML and SELECT queries.
When sending a list of parameter to remote nodes, the PostgreSQL
protocol does not allow types with incorrect OIDs or names, in
the case of XC, incorrect names.
In the case of queries called with SPI, some parameter have no
type as they are not used by all the queries involved on the call
The reason of this behavior is that parameter management is centralized
in the SPI itself. In the case of such parameters, the same method as
dropped attributes of relations is used, meaning that a parameter that
has no type will be considered as a NULL entry and so remote node will
not complain about any unexisting type.
This fix is done on top of the existing infrastructure, but honestly
this should be rewritten with a set of APIs that makes remote parameter
management more easily manageable in cluster, however this is out of
scope for the time being.

2 regression tests are fixed thanks to that: plpgsql and rangefuncs.

Conflicts:
	src/backend/nodes/copyfuncs.c
	src/backend/pgxc/plan/planner.c
	src/backend/pgxc/pool/execRemote.c
	src/pl/plpgsql/src/pl_exec.c
	src/test/regress/expected/rangefuncs_1.out