Change current value of sequences to make it consistent with vanilla
In Postgres-XC (since the beginning of the project in 2009?), the current
value of sequences was taken directly from GTM. However, in vanilla postgres
this is maintained at session level, meaning that the current value of a
sequence inside a session is updated based on the cache maintained in it and
this cache is updated only each time nextval is called inside session.
This resulted in having currval being managed not at session level but at
a cluster level.

The "fix" here just consists in removing the GTM code for current value
management of sequences on GTM. By the way, it does not make sense to manage
that in a centralized way on GTM just knowing that currval is related to a
unique session between nodes.

This inconsistent behavior has been reported by Nikhil Sontakke.
I just cleaned up the code at the necessary places...

Conflicts:
	src/backend/access/transam/gtm.c
	src/backend/commands/sequence.c
	src/gtm/client/gtm_client.c
	src/gtm/main/gtm_seq.c
	src/include/access/gtm.h
	src/include/gtm/gtm_client.h