Allow primary node data modification with ALTER NODE
When trying to use ALTER NODE on a primary node, node manager was complaining
about the fact that there cannot be two primary nodes even if the data altered
was something else like host IP or port number.
The origin of the error was the error handling of primary node inside node
manager. For a newly-created node, it is checked with this new node is primary
and an error is returned if there is already a node defined. For an altered
node, it is checked if a node different than the one altered is primary and
if yes an error is returned. This check was originally moved inside
check_node_options but this was a bad bet as it decentralized the primary
node error check which is really different for new and altered nodes.