Ensure pooler information is reloaded in active sessions after pgxc_pool_reload() is called
InitMultinodeExecutor() takes care of reloading the datanode and coordinator
handles after cluster configuration changes and user calls pgxc_pool_reload().
This was only being done at the start of a session or in the session which
directly calls pgxc_pool_reload(). But this is not enough for other existing
backend connections. We fix that by calling InitMultinodeExecutor() while
handling the pooler invalidation interrupt message